OOTD/code/requirements.txt
asgiref==3.2.3
astroid==2.3.3
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
colorthief==0.2.1
dj-database-url==0.5.0
Django==2.2.7
django-appconf==1.0.3
django-bootstrap4==1.0.1
django-extensions==2.2.5
django-heroku==0.3.1
django-imagekit==4.0.2
gunicorn==20.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pilkit==2.0
Pillow==6.2.1
psycopg2==2.8.4
pylint==2.4.4
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
whitenoise==5.0
wrapt==1.11.2
